copy

+++ audit/src/main/java/com/lovo/audit/dao/lh/ISpecificationDao.java
@@ -1,12 +1,12 @@
 package com.lovo.audit.dao.lh;
 
-import com.lovo.audit.entity.lh.SpecificationEntity;
+import com.lovo.audit.entity.lxj.SpecificationEntity;
 import org.springframework.data.repository.CrudRepository;
 
 /**
- * 促销商品规格持久类
+ * 商品规格持久类
  */
-public interface ISpecificationDao extends CrudRepository<SpecificationEntity,Long> {
+public interface ISpecificationDao extends CrudRepository<SpecificationEntity,Integer> {
 
 
 }
